// apps/api/src/copart/copart.service.d.ts
export declare class CopartService {
    private prisma;
    private titleMapping;
    constructor(prisma: PrismaService, titleMapping: TitleMappingService);
    findAll(query: QueryCopartDto): Promise<{
        data: {
            lotNumber: string;
            inspectable: boolean;
            trim: string | null;
            createdAt: Date;
            updatedAt: Date;
            vehicleType: string | null;
            fuelType: string | null;
            vin: string | null;
            engine: string | null;
            cylinders: string | null;
            odometerBrand: string | null;
            yardName: string | null;
            yardNumber: number | null;
// ...
        }[];
        meta: {
            page: number;
            limit: number;
            total: number;
            totalPages: number;
            hasNextPage: boolean;
            hasPreviousPage: boolean;
        };
    }>;
// ...
}